Ensure that the Date: header is written
The client may have a skewed clock, so make sure that we send out a
"Date:" header. The client will then resend requests using this value.
As a side-effect, convert the main display to show the times as GMT
rather than UTC.

Convert the display to use gmtime()
update the tests